Add story tags attribute
This is a forthcoming change in Twine 2.4. An early form of this functionality has landed on [the develop branch](https://github.com/klembot/twinejs/tree/develop) if you'd like to take a look, but keep in mind that the code is in pre-alpha state and could easily trash any existing stories/cause weird issues if you downgrade to 2.3.